pyproject.toml
[project]
name = "derive-flows"
version = "0.1.0"
description = "Telegram trade message ingestion for options flow analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.34.0",
    "pandas>=2.1.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
derive-flows = "derive_flows.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/derive_flows"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
